The Great Flattening: moved everything from under the dsdl_parser sub-package because it's getting too interdependent with the rest of the library.
