--- pydsdl/dsdl_parser/__init__.py
@@ -1,7 +1,0 @@
-#
-# Copyright (C) 2018-2019  UAVCAN Development Team  <uavcan.org>
-# This software is distributed under the terms of the MIT License.
-#
-
-from .parser import parse_definition, ConfigurationOptions, PrintDirectiveOutputHandler
-from .parser import SemanticError, DSDLSyntaxError, AssertionCheckFailureError, UndefinedDataTypeError
